On this commit I added the postman collection as a file and also on the readme.md file.
On the readme.md file I added all the table and stored procedure queries need in this project.

+++ src/main/java/com/sinzore/benjamin/bankingtestapplication/models/requestbody/administration/AdminLoginModel.java
@@ -4,11 +4,13 @@ import lombok.Data;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigInteger;
+
 @Getter
 @Setter
 @Data
 public class AdminLoginModel {
 
-    int employeeID;
+    BigInteger employeeID;
     String employeePassword;
 }
