+++ src/main/java/com/sinzore/benjamin/bankingtestapplication/models/requestbody/administration/ValidateCustomerAccModel.java
@@ -4,10 +4,12 @@ import lombok.Data;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigInteger;
+
 @Getter
 @Setter
 @Data
 public class ValidateCustomerAccModel {
 
-    int accountNumber;
+    BigInteger accountNumber;
 }
